Kevin

+++ controllers/default.py
@@ -120,16 +120,18 @@ def api():
     def POST(*args, **vars):
         uid = args[0]
         dFlag = args[1]
+        if dFlag == "2":
+            db.users.insert(username=uid, firstName=vars['firstName'])
+            return
         s = vars["dayStart"]
         t = vars["dayEnd"]
-        duration = []
+        duration = []        
         for i in range (int(s), int(t)+1):
           duration.append(i)
         if dFlag == "1":
             db(db.events.startTime==vars['timeStart'])(db.events.endTime==vars['timeEnd'])(db.events.days==duration).delete()
         if dFlag == "0":
-            db.events.insert(username =uid, startTime = vars['timeStart'], endTime = vars['timeEnd'], days = duration)
-        return "store data successfully"
+            db.events.insert(username=uid, startTime = vars['timeStart'], endTime = vars['timeEnd'], days = duration)
 
     def PUT(table_name,record_id,**vars):
         return db(db[table_name]._id==record_id).update(**vars)
